Move signals of user_management to signals.py
