+++ euth/user_management/apps.py
@@ -2,4 +2,8 @@ from django.apps import AppConfig
 
 
 class UsersConfig(AppConfig):
-    name = 'user_management'
+    name = 'euth.user_management'
+    label = 'user_management'
+
+    def ready(self):
+        import euth.projects.signals  # noqa:F401
